+++ src/main/java/net/geforcemods/securitycraft/screen/SetPasswordScreen.java
@@ -11,6 +11,7 @@ import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.screen.Screen;
 import net.minecraft.client.gui.widget.TextFieldWidget;
 import net.minecraft.client.gui.widget.button.Button;
+import net.minecraft.client.util.InputMappings;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.text.IFormattableTextComponent;
@@ -82,6 +83,21 @@ public class SetPasswordScreen extends Screen {
 			font.draw(matrix, blockName, width / 2 - font.width(blockName) / 2, topPos + 6, 4210752);
 			font.draw(matrix, setup, width / 2 - font.width(setup) / 2, topPos + 16, 4210752);
 		}
+	}
+
+	@Override
+	public boolean keyPressed(int keyCode, int scanCode, int modifiers) {
+		if (minecraft.options.keyInventory.isActiveAndMatches(InputMappings.getKey(keyCode, scanCode))) {
+			onClose();
+			return true;
+		}
+
+		return super.keyPressed(keyCode, scanCode, modifiers);
+	}
+
+	@Override
+	public boolean isPauseScreen() {
+		return false;
 	}
 
 	private void saveAndContinueButtonClicked(Button button) {